# internal/app/search/handler.go
package search

import (
	"net/http"
	"sort"
	"strconv"

	"github.com/albertchriss/Tubes2_BE_stami/internal/app/socket"
	"github.com/albertchriss/Tubes2_BE_stami/internal/core"
	"github.com/albertchriss/Tubes2_BE_stami/internal/scraper"
	"github.com/gin-gonic/gin"
)

type ElementResponse struct {
	Value string `json:"value"`
	Label string `json:"label"`
}

type SearchResponse struct {
	Message string           `json:"message"`
	Result  scraper.TreeNode `json:"result"`
}

type BidirectionalSearchResponse struct {
	Message         string                       `json:"message"`
	Result          scraper.BidirectionalResult  `json:"result"`
	// NodesVisited    int                          `json:"nodes_visited"`
	// TimeTaken       string                       `json:"time_taken"`
}

type Handler struct {
	service   Service
	AppCtx    *core.AppContext
  wsManager *socket.ClientManager
}

func NewHandler(service Service, appCtx *core.AppContext, wsManager *socket.ClientManager) *Handler {
	return &Handler{
		service:   service,
		AppCtx:    appCtx,
    wsManager: wsManager,
	}
}

// BFSSearchHandler godoc
// @Summary BFS search handler
// @Description Search the recipe of elements using BFS
// @Tags Search
// @Accept json
// @Produce json
// @Param q query string true "Query parameter"
// @Param num query string false "Number of recipes to return" default(1)
// @Param live query string false "Live update" default(false)
// @Success 200 {object} SearchResponse
// @Router /search/bfs [get]
func (h *Handler) BFSSearchHandler(c *gin.Context) {
	query := c.Query("q")
	if query == "" {
		c.JSON(http.StatusBadRequest, SearchResponse{
			Message: "Query parameter is required",
		})
		return
	}

	numRecipe := c.DefaultQuery("num", "1")
	numRecipeInt, err := strconv.Atoi(numRecipe)
	if err != nil {
		c.JSON(http.StatusBadRequest, SearchResponse{
			Message: "num parameter must be an integer",
		})
		return
	}

	if numRecipeInt < 1 {
		c.JSON(http.StatusBadRequest, SearchResponse{
			Message: "num parameter must be greater than 0",
		})
		return
	}

	liveUpdate := c.DefaultQuery("live", "false")
	liveUpdateBool, err := strconv.ParseBool(liveUpdate)
	if err != nil {
		c.JSON(http.StatusBadRequest, SearchResponse{
			Message: "live parameter must be a boolean",
		})
		return
	}

	res := h.service.BFSSearch(query, numRecipeInt, liveUpdateBool)
	c.JSON(http.StatusOK, SearchResponse{
		Message: "BFS search completed",
		Result:  res,
	})
}

// DFSSearchHandler godoc
// @Summary DFS search handler
// @Description Search the recipe of elements using DFS
// @Tags Search
// @Accept json
// @Produce json
// @Param q query string true "Query parameter"
// @Param num query string false "Number of recipes to return" default(1)
// @Param live query string false "Live update" default(false)
// @Success 200 {object} SearchResponse
// @Router /search/dfs [get]
func (h *Handler) DFSSearchHandler(c *gin.Context) {
	query := c.Query("q")
	if query == "" {
		c.JSON(http.StatusBadRequest, SearchResponse{
			Message: "Query parameter is required",
		})
		return
	}

	numRecipe := c.DefaultQuery("num", "1")
	numRecipeInt, err := strconv.Atoi(numRecipe)
	if err != nil {
		c.JSON(http.StatusBadRequest, SearchResponse{
			Message: "num parameter must be an integer",
		})
		return
	}

	if numRecipeInt < 1 {
		c.JSON(http.StatusBadRequest, SearchResponse{
			Message: "num parameter must be greater than 0",
		})
		return
	}

	liveUpdate := c.DefaultQuery("live", "false")
	liveUpdateBool, err := strconv.ParseBool(liveUpdate)
	if err != nil {
		c.JSON(http.StatusBadRequest, SearchResponse{
			Message: "live parameter must be a boolean",
		})
		return
	}

	res := h.service.DFSSearch(query, numRecipeInt, liveUpdateBool)
	c.JSON(http.StatusOK, SearchResponse{
		Message: "DFS search completed",
		Result:  res,
	})
}

// BidirectionalSearchHandler godoc
// @Summary Bidirectional search handler
// @Description Search the recipe of elements using Bidirectional Search. Returns two trees representing the search paths from base elements and from the target element, meeting at a common node.
// @Tags Search
// @Accept json
// @Produce json
// @Param q query string true "Target element to search for"
// @Success 200 {object} BidirectionalSearchResponse "Successful search operation"
// @Router /search/bidirectional [get]
func (h *Handler) BidirectionalSearchHandler(c *gin.Context) {
	query := c.Query("q")
	if query == "" {
		c.JSON(http.StatusBadRequest, BidirectionalSearchResponse{
			Message: "Query parameter is required",
		})
		return
	}

	res := h.service.BidirectionalSearch(query)
	c.JSON(http.StatusOK, BidirectionalSearchResponse{
		Message: "Bidirectional search completed",
		Result:  res,
	})
}

// GetElementsHandler godoc
// @Summary Get all elements
// @Description Get a list of all available elements
// @Tags Elements
// @Produce json
// @Success 200 {array} ElementResponse
// @Router /elements [get]
func (h *Handler) GetElementsHandler(c *gin.Context) {
	tierData := h.AppCtx.Config.TierMap
	if tierData == nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Tier data not loaded"})
		return
	}

	var elements []ElementResponse
	for elementName := range *tierData {
		elements = append(elements, ElementResponse{Value: elementName, Label: elementName})
	}

	sort.Slice(elements, func(i, j int) bool {
		return elements[i].Label < elements[j].Label
	})

	c.JSON(http.StatusOK, elements)
}